feat: Move the cards and savedCards arrays from ./src/components/Movies/MoviesCardList/MoviesCardList.js to ./src/utils/constants.js
