+++ src/components/Movies/MoviesCardList/MoviesCardList.js
@@ -4,113 +4,10 @@ import useWindowResize from '../../../hooks/useWindowResize';
 
 import './MoviesCardList.css';
 
-import movieImage from '../../../images/movie-image.jpg';
+import { cards, savedCards } from '../../../utils/constants'
 
 import Preloader from '../Preloader/Preloader';
 import MoviesCard from '../MoviesCard/MoviesCard';
-
-// Временный массив карточек, без использования API
-const cards = [
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-]
-
-// Временный массив сохранненых карточек, без использования API
-const savedCards = [
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-  {
-    image: movieImage,
-    name: '33 слова о дизайне',
-    duration: '1ч42м',
-  },
-]
 
 function MoviesCardList() {
