+++ src/utils/constants.js
@@ -1,0 +1,106 @@
+import movieImage from '../images/movie-image.jpg';
+
+// Временный массив карточек, без использования API
+const cards = [
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+]
+
+// Временный массив сохранненых карточек, без использования API
+const savedCards = [
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+  {
+    image: movieImage,
+    name: '33 слова о дизайне',
+    duration: '1ч42м',
+  },
+]
+
+export { cards, savedCards };
